+++ src/localization/i18n.ts
@@ -1,5 +1,5 @@
 import VueI18n from 'vue-i18n'
-import { localizations } from '@/localization/localizations/en'
+import { localizations } from '@/localization/localizations'
 import Vue from 'vue'
 
 Vue.use(VueI18n)
